requirements.txt
python-telegram-bot==21.0.1
requests==2.31.0
openai==1.3.0
pytesseract==0.3.10
Pillow==10.1.0
python-dotenv==1.0.0
pydantic==2.5.0
aiohttp==3.9.1
